requirements.txt
transformers>=4.36.2
diffusers>=0.25.0
ctransformers[cuda]>=0.2.24
Pillow>=9.3.0
streamlit>=1.29.0
accelerate>=0.25.0
pypdf
peft>=0.6.0
unstructured==0.9.2
tiktoken==0.4.0
faiss-cpu==1.7.4
libmagic==1.0
python-magic==0.4.27
python-magic-bin==0.4.14
dotenv
langchain
langchain-community
openai
requests
io
youtube_transcript_api
newspaper

torch